+++ tests/Functional/Rest/Credentials.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Test\Functional\Rest;
+
+class Credentials
+{
+    public static function getAdminUser()
+    {
+        return [
+            'username' => (getenv('TEST_ADMIN_USER') ? getenv('TEST_ADMIN_USER') : 'admin'),
+            'password' => (getenv('TEST_ADMIN_PASSWORD') ? getenv('TEST_ADMIN_PASSWORD') : 'pwd'),
+        ];
+    }
+
+    public static function getRegularUser()
+    {
+        return [
+            'username' => (getenv('TEST_REGULAR_USER') ? getenv('TEST_REGULAR_USER') : 'user'),
+            'password' => (getenv('TEST_REGULAR_PASSWORD') ? getenv('TEST_REGULAR_PASSWORD') : 'pwd'),
+        ];
+    }
+}
